refactor code to monitoringapi file
